requirements.txt
streamlit
google-generativeai
requests
beautifulsoup4
lxml
pandas
python-dotenv 